+++ src/pages/_document.tsx
@@ -7,7 +7,6 @@ import Document, {
   DocumentInitialProps,
 } from 'next/document';
 import React from 'react';
-import { gaMeasurementId } from '../infrastructures/utils/gtag';
 
 export default class CustomDocument extends Document {
   static async getInitialProps(
@@ -23,26 +22,7 @@ export default class CustomDocument extends Document {
   render(): JSX.Element {
     return (
       <Html prefix="og: https://ogp.me/ns#">
-        <Head>
-          {/* Google Analytics */}
-          <script
-            async
-            src={`https://www.googletagmanager.com/gtag/js?id=${gaMeasurementId}`}
-          />
-          <script
-            // eslint-disable-next-line react/no-danger
-            dangerouslySetInnerHTML={{
-              __html: `
-                window.dataLayer = window.dataLayer || [];
-                function gtag(){dataLayer.push(arguments);}
-                gtag('js', new Date());
-                gtag('config', '${gaMeasurementId}', {
-                  page_path: window.location.pathname,
-                });
-              `,
-            }}
-          />
-        </Head>
+        <Head />
         <body>
           <Main />
           <NextScript />
